requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
flask-migrate
passlib
gunicorn
psycopg2-binary
requests
rq
redis